+++ src/glut_compatability.cxx
@@ -481,7 +481,7 @@ int glutExtensionSupported( const char* extension )
   if (!extension || strchr(extension, ' ')) return 0;
 
   const char *extensions, *start;
-  const int len = strlen( extension );
+  const int len = static_cast<int>(strlen( extension ));
   
   start = extensions = (const char *) glGetString(GL_EXTENSIONS);
